setup_model.py and setup_optimizer.py moved to seprate files and added type hinits
